requirements.txt
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0
requests>=2.31.0
pyarrow>=15.0.0
xgboost>=2.0.0
scikit-learn>=1.4.0
torch>=2.1.0
plotly>=5.18.0
streamlit>=1.31.0
holidays>=0.40
shap>=0.44.0
statsmodels>=0.14.0
